fix codex

+++ services/dbms/src/distributed_bms.cpp
@@ -334,7 +334,8 @@ std::unique_ptr<char[]> DistributedBms::EncodeBase64(std::unique_ptr<unsigned ch
             DECODE_TABLE[(static_cast<uint8_t>(byte1 & DECODE_VALUE_CHAR_THREE) << DECODE_VALUE_FOUR)
              | (byte2 >> DECODE_VALUE_FOUR)];
         dstData[j++] =
-            DECODE_TABLE[((byte2 & DECODE_VALUE_CHAR_FIFTEEN) << DECODE_VALUE_TWO) | (byte3 >> DECODE_VALUE_SIX)];
+            DECODE_TABLE[(static_cast<uint8_t>(byte2 & DECODE_VALUE_CHAR_FIFTEEN)
+                << DECODE_VALUE_TWO) | (byte3 >> DECODE_VALUE_SIX)];
         dstData[j++] = DECODE_TABLE[byte3 & DECODE_VALUE_CHAR_SIXTY_THREE];
     }
     if (srcLen % DECODE_VALUE_THREE == DECODE_VALUE_ONE) {
@@ -350,7 +351,8 @@ std::unique_ptr<char[]> DistributedBms::EncodeBase64(std::unique_ptr<unsigned ch
         dstData[j++] =
             DECODE_TABLE[(static_cast<uint8_t>(byte1 & DECODE_VALUE_CHAR_THREE) << DECODE_VALUE_FOUR)
              | (byte2 >> DECODE_VALUE_FOUR)];
-        dstData[j++] = DECODE_TABLE[(byte2 & DECODE_VALUE_CHAR_FIFTEEN) << DECODE_VALUE_TWO];
+        dstData[j++] = DECODE_TABLE[static_cast<uint8_t>(byte2 & DECODE_VALUE_CHAR_FIFTEEN)
+                                    << DECODE_VALUE_TWO];
         dstData[j++] = '=';
     }
     dstData[outLen] = '\0';
